renamed get_mayavi2_engine_manager.py to mayavi2.py
