+++ simphony_mayavi/api.py
@@ -2,8 +2,8 @@ from .show import show
 from .snapshot import snapshot
 from .adapt2cuds import adapt2cuds
 from .load import load
-from .get_mayavi2_engine_manager import add_engine_to_mayavi2
-from .get_mayavi2_engine_manager import get_simphony_panel
+from .mayavi2 import add_engine_to_mayavi2
+from .mayavi2 import get_simphony_panel
 from .restore_scene import restore_scene
 
 __all__ = ['show', 'snapshot', 'adapt2cuds', 'load',
